requirements.txt
streamlit>=1.38.0
openai>=1.51.0
python-dotenv>=1.0.1